Debug Inode ctor/dctor; fix printf args warnigns
git-svn-id: file:///srv/svn/repos/haiku/trunk/current@6285 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/kernel/file_systems/bfs/Inode.h
@@ -383,9 +383,9 @@ inline void
 CachedBlock::KDumpMe()
 {
 	kprintf("CachedBlock {\n");
-	kprintf("fVolume = 0x%08lx\n", fVolume);
+	kprintf("fVolume = 0x%08lx\n", (uint32)fVolume);
 	kprintf("fBlockNumber = 0x%16Lx\n", fBlockNumber);
-	kprintf("fBlock = 0x%08lx\n", fBlock);
+	kprintf("fBlock = 0x%08lx\n", (uint32)fBlock);
 	kprintf("}\n");
 }
